requirements.txt
a2a-sdk>=0.2.5
httpx>=0.28
uvicorn>=0.34
